Fix how the data is being stored into the packet class
Had an issue with putting the contents of data as the FileID and packetNumber,
which is wrong because data is just the content of the received packet. So instead
it stores it into a temp array, and the temp array contents get distributed into the
correct fields. Also added a way for it to get sent to a File.

Now the issue is how we wait for all the packets to be received. Technically our math is wrong, because
there are multiple files being sent so we don't know when to stop at the moment. Currently we
wait for a end packet and then wait for enough packets to be received to meet that total, but since there
are three files, there are also three end packets. Therefore, we need to change how we wait for the end.